requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-docx==1.1.0
python-pptx==1.0.2
PyMuPDF==1.26.3
docx2txt==0.9
txt2docx== 1.0.2
reportlab==4.4.3
pypandoc==1.11
crewai==0.159.0
langchain==0.1.0
langchain-openai==0.0.2
openai==1.3.7
whisper==1.1.10
pyttsx3==2.90
Markdown 3.8.2
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.0 
